resolved search results bug
