+++ public/js/profile.js
@@ -4,13 +4,13 @@ const dbSearchInput = document.getElementById("db-search");
 dbSearchBtn.addEventListener("click", (event) => {
     event.preventDefault();
     console.log("DB Search Btn Hit");
-    const dbSearchValue = { "dbSearchTerm": dbSearchInput.value.trim() };
+    const dbSearchValue = dbSearchInput.value.trim();
     console.log(dbSearchValue)
 
-    fetch("/api/db_search", { method: "POST", body: JSON.stringify(dbSearchValue), headers: { 'Content-Type': 'application/json' } })
+    fetch(`/api/search-results/${dbSearchValue}`)
         .then(() => {
             console.log("Search Submitted")
-            window.location.replace(`/search-results`);
+            window.location.replace(`/search-results/${dbSearchValue}`);
         })
         .catch(err => console.log(err))
 });
